Pre-CLI commit

+++ CrudexBackend/CrudexBackend/Models/Nature.cs
@@ -7,10 +7,10 @@ namespace CrudexBackend.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int AttackMod { get; set; }
-        public int DefenseMod { get; set; }
-        public int SpAttackMod { get; set; }
-        public int SpDefenseMod { get; set; }
-        public int SpeedMod { get; set; }
+        public double AttackMod { get; set; }
+        public double DefenseMod { get; set; }
+        public double SpAttackMod { get; set; }
+        public double SpDefenseMod { get; set; }
+        public double SpeedMod { get; set; }
     }
 }
